requirements.txt
pandas==2.1.4
reportlab==4.0.8
streamlit==1.30.0
PyPDF2==1.26.0
reportlab==4.0.9
pillow==10.2.0
protobuf>=3.20,<5
altair==4.1.0